+++ musique/interpreter/interpreter.hh
@@ -6,6 +6,13 @@
 #include <musique/midi/midi.hh>
 #include <musique/value/value.hh>
 #include <unordered_map>
+#include <set>
+
+struct KeyboardInterrupt : std::exception
+{
+	~KeyboardInterrupt() = default;
+	char const* what() const noexcept override { return "KeyboardInterrupt"; }
+};
 
 /// Given program tree evaluates it into Value
 struct Interpreter
@@ -21,6 +28,8 @@ struct Interpreter
 	std::shared_ptr<Context> current_context;
 
 	std::function<std::optional<Error>(Interpreter&, Value)> default_action;
+
+	std::multiset<std::pair<unsigned, unsigned>> active_notes;
 
 	Starter starter;
 
@@ -53,6 +62,18 @@ struct Interpreter
 
 	/// Dumps snapshot of interpreter into stream
 	void snapshot(std::ostream& out);
+
+	/// Turn all notes that have been played but don't finished playing
+	void turn_off_all_active_notes();
+
+	/// Handles interrupt if any occured
+	void handle_potential_interrupt();
+
+	/// Issue new interrupt
+	void issue_interrupt();
+
+	/// Sleep for at least given time or until interrupt
+	void sleep(std::chrono::duration<float>);
 };
 
 std::optional<Error> ensure_midi_connection_available(Interpreter&, std::string_view operation_name);